+++ test/spec_mock_request.rb
@@ -5,7 +5,7 @@ require 'yaml'
 require_relative 'psych_fix'
 
 separate_testing do
-  require_relative '../lib/rack/mock'
+  require_relative '../lib/rack/mock_request'
   require_relative '../lib/rack/lint'
   require_relative '../lib/rack/request'
   require_relative '../lib/rack/response'
